# openapi/openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/onramp/quotes:
    post:
      operationId: OnrampController_createQuote_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePayinQuoteDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayinQuoteEntity"
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key lacks the required scope, or the request did not arrive
            through the gateway.
          content:
            application/json:
              schema: *a1
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Conflicts with existing state — an idempotency clash, an operation
            already in flight, or an invalid state transition.
          content:
            application/json:
              schema: *a1
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: Create a payin quote (expires in ~5 min)
      tags: &a17
        - onramp
  /v1/onramp/payins:
    post:
      operationId: OnrampController_createPayin_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePayinDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayinEntity"
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key lacks the required scope, or the request did not arrive
            through the gateway.
          content:
            application/json:
              schema: *a1
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Conflicts with existing state — an idempotency clash, an operation
            already in flight, or an invalid state transition.
          content:
            application/json:
              schema: *a1
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: Create a payin from a quote; returns funding instructions
      tags: *a17
    get:
      operationId: OnrampController_findAll_v1
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayinListEntity"
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key lacks the required scope, or the request did not arrive
            through the gateway.
          content:
            application/json:
              schema: *a1
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Conflicts with existing state — an idempotency clash, an operation
            already in flight, or an invalid state transition.
          content:
            application/json:
              schema: *a1
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: List the consumer's payins
      tags: *a17
  /v1/onramp/payins/{id}:
    get:
      operationId: OnrampController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayinEntity"
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key lacks the required scope, or the request did not arrive
            through the gateway.
          content:
            application/json:
              schema: *a1
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Conflicts with existing state — an idempotency clash, an operation
            already in flight, or an invalid state transition.
          content:
            application/json:
              schema: *a1
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: Get a payin (serves the local mirror; refreshes from BlindPay when stale)
      tags: *a17
  /v1/onramp/trustline:
    post:
      operationId: OnrampController_createTrustline_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrustlineDto"
      responses:
        "201":
          description: ""
        "400":
          description: Validation failed, or the request is not valid in the current state.
          content:
            application/json:
              schema: *a1
        "401":
          description: No authenticated consumer, or admin credentials are required.
          content:
            application/json:
              schema: *a1
        "403":
          description: The API key lacks the required scope, or the request did not arrive
            through the gateway.
          content:
            application/json:
              schema: *a1
        "404":
          description: The resource does not exist, or does not belong to this consumer.
          content:
            application/json:
              schema: *a1
        "409":
          description: Conflicts with existing state — an idempotency clash, an operation
            already in flight, or an invalid state transition.
          content:
            application/json:
              schema: *a1
        "500":
          description: Unexpected server error. The detail is logged server-side and never
            returned.
          content:
            application/json:
              schema: *a1
        "502":
          description: An upstream provider returned an error.
          content:
            application/json:
              schema: *a1
        "503":
          description: An upstream provider is unavailable, or the service is misconfigured.
          content:
            application/json:
              schema: *a1
      summary: Build an unsigned Stellar trustline tx (XDR) for the customer to sign
      tags: *a17
...
info:
  title: Cosmos Pay — Payments API
  description: "Payments microservice (Stellar payment intents). All endpoints
    require traffic to arrive through the APISIX gateway: a valid
    `X-Gateway-Secret` header plus an authenticated consumer
    (`X-Consumer-Username`). Paths already include the version (`/v1/...`)."
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    gateway-secret:
      type: apiKey
      in: header
      name: X-Gateway-Secret
    consumer:
      type: apiKey
      in: header
      name: X-Consumer-Username
  schemas:
...
    PayerRulesDto:
      type: object
      properties:
        pix_allowed_tax_ids:
          type: array
          items:
            type: string
        transfers_allowed_tax_id:
          type: string
          example: "20123456786"
          description: Argentine payer CUIT/CUIL (Transfers 3.0).
        pse_allowed_tax_ids:
          type: array
          items:
            type: string
        pse_full_name:
          type: string
          example: Jane Doe
        pse_document_type:
          type: string
          example: CC
          description: CC | NIT
        pse_document_number:
          type: string
          example: "123456789"
        pse_email:
          type: string
          example: jane@acme.com
        pse_phone:
          type: string
          example: "+573001234567"
        pse_bank_code:
          type: string
          example: "9"
    CreatePayinQuoteDto:
      type: object
      properties:
        blockchain_wallet_id:
          type: string
          example: clz9xwallet001
          description: Cosmos Pay wallet id (from POST /v1/kyc/receivers/:id/wallets) that
            receives the minted stablecoin.
        currency_type:
          type: string
          enum: &a26
            - sender
            - receiver
          example: sender
        payment_method:
          type: string
          enum:
            - ach
            - wire
            - pix
            - ted
            - spei
            - transfers
            - pse
            - international_swift
            - rtp
          example: pix
          description: "Fiat-in method. Note: payins use `spei`/`transfers`/`pse` (not the
            `_bitso` payout rails)."
        token:
          type: string
          enum: &a25
            - USDC
            - USDT
            - USDB
          example: USDC
        request_amount:
          type: number
          example: 10000
          description: Amount in minor units (integer).
        cover_fees:
          type: boolean
          example: false
          description: When true, the sender absorbs the fees.
        payer_rules:
          description: "Required for Transfers (AR): transfers_allowed_tax_id. Required
            for PSE (CO): pse_full_name, pse_document_type, pse_document_number,
            pse_email, pse_bank_code. Optional for other methods (e.g. pix)."
          allOf:
            - $ref: "#/components/schemas/PayerRulesDto"
        is_otc:
          type: boolean
          example: false
        partner_fee_id:
          type: string
        wallet_id:
          type: string
          description: Optional custodial wallet id.
      required:
        - blockchain_wallet_id
        - currency_type
        - payment_method
        - token
        - request_amount
    PayinQuoteEntity:
      type: object
      properties:
        id:
          type: string
          example: pi_000000000000
        expires_at:
          type: number
          example: 1782692400
          description: Unix expiry (~5 min out).
        commercial_quotation:
          type: number
          example: 5.42
        receiver_amount:
          type: number
          example: 10000
        sender_amount:
          type: number
          example: 10000
      required:
        - id
        - expires_at
    CreatePayinDto:
      type: object
      properties:
        payin_quote_id:
          type: string
          example: qu_000000000000
          description: The id returned by POST /v1/onramp/quotes.
      required:
        - payin_quote_id
    PayinEntity:
      type: object
      properties:
        id:
          type: string
          example: clz9xpayin0001
        blindpayId:
          type: string
          example: pi_000000000000
        status:
          type: object
          example: processing
        token:
          type: object
          example: USDC
        network:
          type: object
          example: stellar
        paymentMethod:
          type: object
          example: pix
        senderAmount:
          type: object
          example: "10000"
        receiverAmount:
          type: object
          example: "9950"
        instructions:
          type: object
          description: Funding instructions for the payer (memo/CLABE/PIX/CBU/PSE/...).
          example:
            pix_code: 00020126...5204000053039865802BR...
        createdAt:
          format: date-time
          type: string
          example: 2026-06-28T12:00:00.000Z
      required:
        - id
        - blindpayId
        - createdAt
    PayinListEntity:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/PayinEntity"
        total:
          type: number
          example: 1
      required:
        - data
        - total
    CreateTrustlineDto:
      type: object
      properties:
        address:
          type: string
          example: GCALNQQBXAPZ2WIRSDDBMSTAKCUH5SG6U76YBFLQLIXJTF7FE5AX7AOO
          description: Stellar address that needs the asset trustline.
      required:
        - address
...
security:
  - gateway-secret: []
  - consumer: []